Split out altair file and add increase tracing
