+++ ibis_vega_transform/__init__.py
@@ -5,8 +5,15 @@ evaluated with ibis expressions.
 The main function is the ``ibis_vega_transform.apply()`` function.
 """
 from .core import apply
-from . import transforms
-from .altair import display_queries
+from . import (
+    transforms,
+    altair_data_transformer,
+    altair_monkeypatch,
+    altair_renderer,
+    compiler,
+    query,
+)
+from .query import display_queries
 
 
 __all__ = ["apply", "transforms", "display_queries"]
