+++ src/renderer.ts
@@ -26,28 +26,43 @@ export class IbisVegaRenderer extends Widget implements IRenderMime.IRenderer {
    * Render vega data using the alternative renderer.
    */
   async renderModel(model: IRenderMime.IMimeModel): Promise<void> {
-    const vlSpec = model.data[MIME_TYPE] as any;
+    const {spec: vlSpec, span} = model.data[MIME_TYPE] as {spec: any, span: any};
     const kernel = this._context.session.kernel;
 
     if (kernel === null) {
       return;
     }
+
+    // Skip rendering if we are already rendering or we are about to rerender the same spec
+    if (
+      this._renderingSpec ||
+      // TODO: Replace this with real deep equality
+      (this._renderedSpec &&
+        JSON.stringify(this._renderedSpec) === JSON.stringify(vlSpec))
+    ) {
+      return;
+    }
+    this._renderingSpec = true;
+
     if (this._view) {
       this._view.finalize();
       this._view = null;
     }
 
+    const vSpec = await compileSpec(kernel, vlSpec, span);
+
     // TODO: is this safe if there are multiple kernels trying to use
     // the transform? Can we construct individual transforms/comms for
     // each renderer? What does IPywidgets do?
     ibisTransform.kernel = kernel;
-    const vSpec = await compileSpec(kernel, vlSpec);
     const res = await vegaEmbed(this.node, vSpec, {
       actions: true,
       defaultStyle: true,
       mode: 'vega'
     });
     this._view = res.view;
+    this._renderingSpec = false;
+    this._renderedSpec = vlSpec;
   }
 
   /**
@@ -72,4 +87,6 @@ export class IbisVegaRenderer extends Widget implements IRenderMime.IRenderer {
 
   private _isDisposed = false;
   private _view: vega.View | null = null;
+  private _renderingSpec = false;
+  private _renderedSpec: any = null;
 }
